+++ cmd/captplanet/run.go
@@ -12,24 +12,9 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/pkg/accounting/rollup"
-	"storj.io/storj/pkg/accounting/tally"
-	"storj.io/storj/pkg/audit"
 	"storj.io/storj/pkg/auth/grpcauth"
-	"storj.io/storj/pkg/bwagreement"
 	"storj.io/storj/pkg/cfgstruct"
-	"storj.io/storj/pkg/datarepair/checker"
-	"storj.io/storj/pkg/datarepair/repairer"
-	"storj.io/storj/pkg/discovery"
-	"storj.io/storj/pkg/inspector"
-	"storj.io/storj/pkg/kademlia"
-	"storj.io/storj/pkg/miniogw"
-	"storj.io/storj/pkg/overlay"
-	"storj.io/storj/pkg/piecestore/psserver"
-	"storj.io/storj/pkg/pointerdb"
 	"storj.io/storj/pkg/process"
-	"storj.io/storj/pkg/satellite/satelliteweb"
-	"storj.io/storj/pkg/server"
 	"storj.io/storj/pkg/utils"
 	"storj.io/storj/satellite/satellitedb"
 )
@@ -38,31 +23,6 @@ const (
 	storagenodeCount = 10
 )
 
-// Satellite is for configuring client
-type Satellite struct {
-	Server      server.Config
-	Kademlia    kademlia.SatelliteConfig
-	PointerDB   pointerdb.Config
-	Overlay     overlay.Config
-	Inspector   inspector.Config
-	Checker     checker.Config
-	Repairer    repairer.Config
-	Audit       audit.Config
-	BwAgreement bwagreement.Config
-	Web         satelliteweb.Config
-	Database    string `help:"satellite database connection string" default:"sqlite3://$CONFDIR/master.db"`
-	Discovery   discovery.Config
-	Tally       tally.Config
-	Rollup      rollup.Config
-}
-
-// StorageNode is for configuring storage nodes
-type StorageNode struct {
-	Server   server.Config
-	Kademlia kademlia.StorageNodeConfig
-	Storage  psserver.Config
-}
-
 var (
 	runCmd = &cobra.Command{
 		Use:   "run",
@@ -70,11 +30,7 @@ var (
 		RunE:  cmdRun,
 	}
 
-	runCfg struct {
-		Satellite    Satellite
-		StorageNodes [storagenodeCount]StorageNode
-		Uplink       miniogw.Config
-	}
+	runCfg Config
 )
 
 func init() {
